Updated Tabs to 4 spaces
Updated .vimrc formatting with headings for plugin section and normal
vim settings.  Changed the tabs to 4 spaces and the default indentation
to 4 spaces;  decided to change because different applications display
tabs differently, and it messes up the visual formatting I tried to
create with a file.  (a file looks good in vim, but gets messed up when
printed out on the terminal)