Added Fetch Installed Packages Script
Created a script that fetches the explicitly installed foreign and
non-foreign (includes yay pkgs) and stores this data in a list;  the
script filters against an ignore packages file.

A pre-commit git hook was created in git_hooks/.  This hook will
update the package lists prior to committing.  These scripts must be
symlinked to .git/hooks/