Added pre-commit framework to track packages
The pre-commit framework is a better way for interacting with git hooks
because abortion of commits when files in staging area are modified is
built-in to the framework.  This ensures that everytime I commit to this
repo, the update package list script will be ran and the changes to the
package lists will be included in the commit.

Changed some variable names in scripts.  Modified
`fetch_installed_packages_list.sh` input parameters to make it more
flexible.