feat(nvim): enforce 2 tabs and make listchar for tab readable
Remember nvim's `listchar` for tabs is the characters that will be
displayed when an actual '\t' character is written in the file.
It was `>` followed by the number of spaces in a tab, but this
is really confusing especially when reading assembly. Changed
to a nerd-font character that is a sequence of vertical dots.