refactor(screenlayout): seperate concerns for auto script
Refactor the auto script which is now called `update_display_layout.sh`
so that it only does one thing: determine which display layout to update
and update it. The script implementation was refactored to make it
simpler.

Updated formatting and word choice in the note-to-self about creating a
default layout called `README_default_layout.md`.